Added check for EGL errors and more log prints.

+++ VGEngine/Engine/include/engine/game/quadrangleComponent.h
@@ -11,6 +11,6 @@ namespace vg{
 
 	private:
 		std::vector<float> getDefaultVertices();
-		std::vector<GLuint> getDefaultIndices();
+		std::vector<GLushort> getDefaultIndices();
 	};
 }
