+++ VGEngine/Engine/source/game/renderComponent.cpp
@@ -3,7 +3,7 @@
 
 using namespace vg;
 
-RenderComponent::RenderComponent(std::vector<float> vertices, std::vector<GLuint> indices) :Component()
+RenderComponent::RenderComponent(std::vector<float> vertices, std::vector<GLushort> indices) :Component()
 {
 	this->vertices = vertices;
 	this->indices = indices;
